start of mvp
clean code which is unused.
