+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".MainActivity"
+            android:name=".view.activity.MainActivity"
             android:label="@string/app_name"
             android:launchMode="singleTask">
             <intent-filter>
@@ -36,17 +36,14 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".RootChatActivity" />
         <activity android:name=".GradeActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"/>
-        <activity android:name=".StartActivity" />
-        <activity android:name=".ThisGradeActivity" />
-        <activity android:name=".HelpActivity"
+        <activity android:name=".view.activity.HelpActivity"
             android:theme="@style/toolBarActivityTheme" />
         <!-- 可选<activity android:name="com.pgyersdk.activity.FeedbackActivity"/>-->
 
 
-        <receiver android:name=".WiFiReceiver">
+        <receiver android:name=".view.widget.WiFiReceiver">
             <intent-filter>
                 <action android:name="android.net.wifi.STATE_CHANGE" />
             </intent-filter>
