--- app/src/main/java/com/jimbo/myapplication/ThisGradeActivity.java
@@ -1,39 +1,0 @@
-package com.jimbo.myapplication;
-
-import android.content.Intent;
-import android.net.Uri;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.webkit.DownloadListener;
-
-import com.jimbo.myapplication.webview.ProgressWebView;
-
-/**
- *
- * Created by jimbo on 15-12-18.
- */
-public class ThisGradeActivity extends AppCompatActivity {
-
-
-    Toolbar toolbar;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_grade);
-        toolbar.setTitle("成绩查询");
-        setSupportActionBar(toolbar);
-        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                onBackPressed();
-            }
-        });
-
-
-
-    }
-}
