+++ app/src/main/java/com/jimbo/myapplication/view/activity/HelpActivity.java
@@ -1,4 +1,4 @@
-package com.jimbo.myapplication;
+package com.jimbo.myapplication.view.activity;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -8,7 +8,9 @@ import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.webkit.DownloadListener;
 
-import com.jimbo.myapplication.webview.ProgressWebView;
+import com.jimbo.myapplication.Config;
+import com.jimbo.myapplication.R;
+import com.jimbo.myapplication.view.view.ProgressWebView;
 
 /**
  *
