--- app/src/main/res/drawable/button_send.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item android:state_pressed="true" android:drawable="@mipmap/send_btn_pressed"></item>
-    <item android:drawable="@mipmap/send_btn_normal"></item>
-
-</selector>
