--- app/src/main/res/layout/activity_newmain.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/Background"
-    android:orientation="vertical">
-
-
-    <co.mobiwise.library.ProgressLayout
-        android:id="@+id/progressLayout"
-        android:layout_width="match_parent"
-        android:layout_height="100dp"
-        app:autoProgress="true"
-        app:emptyColor="@color/q"
-        app:loadedColor="@color/ContentLayoutRefreshing"
-        app:maxProgress="100">
-    </co.mobiwise.library.ProgressLayout>
-
-    <co.mobiwise.library.ProgressLayout
-        android:id="@+id/progressLayout2"
-        android:layout_width="match_parent"
-        android:layout_height="100dp"
-        app:autoProgress="true"
-        app:emptyColor="@color/w"
-        app:loadedColor="@color/ContentLayoutRefreshing"
-        app:maxProgress="100">
-    </co.mobiwise.library.ProgressLayout>
-
-</LinearLayout>
