--- app/src/main/res/layout/activity_root_chat.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <RelativeLayout
-        android:id="@+id/chat_title"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:layout_alignParentTop="true"
-        android:background="#171717"
-        android:gravity="center">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="小bitch"
-            android:textColor="#ffffff"
-            android:textSize="25sp" />
-    </RelativeLayout>
-
-
-    <ListView
-        android:id="@+id/chat_listview"
-        android:layout_width="match_parent"
-        android:layout_above="@+id/chat_send"
-        android:layout_below="@+id/chat_title"
-        android:dividerHeight="0dip"
-        android:divider="#00000000"
-        android:layout_height="match_parent">
-    </ListView>
-
-    <RelativeLayout
-        android:id="@+id/chat_send"
-        android:layout_width="match_parent"
-        android:layout_height="55dp"
-        android:background="#000000"
-        android:gravity="center"
-        android:layout_alignParentBottom="true">
-
-        <EditText
-            android:id="@+id/chat_edittext_message"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textSize="18sp"
-            android:layout_marginLeft="10dp"
-            android:layout_marginRight="5dp"
-            android:background="@mipmap/login_edit_normal"
-            android:layout_toLeftOf="@+id/chat_button_send"
-            />
-
-        <Button
-            android:id="@+id/chat_button_send"
-            android:text="发送"
-            android:layout_alignParentRight="true"
-            android:layout_width="60dp"
-            android:background="@drawable/button_send"
-            android:layout_height="40dp" />
-
-    </RelativeLayout>
-
-
-</RelativeLayout>
