--- app/src/main/res/layout/activity_start.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-    <TextView
-        android:id="@+id/start_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-    </ScrollView>
-
-</LinearLayout>
