--- app/src/main/res/layout/listview_item_send.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-
-    <LinearLayout
-        android:visibility="gone"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingTop="2dp"
-        android:paddingBottom="2dp"
-        android:orientation="vertical">
-        <TextView
-            android:id="@+id/chat_send_date"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:text="2015-09-16 22:54:10" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:gravity="right"
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/listview_item_send_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
-            android:background="@mipmap/chatto_bg_normal"
-            android:text="你好"
-            android:textSize="18sp" />
-
-        <LinearLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <ImageView
-                android:id="@+id/listview_item_send_image"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
-                android:src="@mipmap/me" />
-
-            <TextView
-                android:id="@+id/listview_item_send_name"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:text="diedie"
-                android:textSize="12sp" />
-        </LinearLayout>
-    </LinearLayout>
-</LinearLayout>
